+++ cli/cli/helpers/engine_manager/engine_manager.go
@@ -2,10 +2,11 @@ package engine_manager
 
 import (
 	"context"
-	"fmt"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_cluster_setting"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/resolved_config"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/kurtosis_engine_rpc_api_bindings"
@@ -22,7 +23,7 @@ import (
 
 const (
 	waitForEngineResponseTimeout = 5 * time.Second
-
+	defaultClusterName = resolved_config.DefaultDockerClusterName
 
 	// --------------------------- Old port parsing constants ------------------------------------
 	// These are the old labels that the API container used to use before 2021-11-15 for declaring its port num protocol
@@ -47,14 +48,29 @@ var objAttrsSchemaPortProtosToDockerPortProtos = map[schema.PortProtocol]string{
 }
 
 type EngineManager struct {
-	kurtosisBackend backend_interface.KurtosisBackend
-	shouldSendMetrics bool
+	kurtosisBackend                           backend_interface.KurtosisBackend
+	shouldSendMetrics                         bool
 	engineServerKurtosisBackendConfigSupplier engine_server_launcher.KurtosisBackendConfigSupplier
 	// Make engine IP, port, and protocol configurable in the future
 }
 
-// TODO the fact that this takes in a clusterName is a temporary hack, until we have proper used-cluster data storage
-func NewEngineManager(clusterName string) (*EngineManager, error) {
+func NewEngineManager() (*EngineManager, error) {
+	clusterSettingStore := kurtosis_cluster_setting.GetKurtosisClusterSettingStore()
+
+	isClusterSet, err := clusterSettingStore.HasClusterSetting()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Failed to check if cluster setting has been set.")
+	}
+	var clusterName string
+	if !isClusterSet {
+		// If the user has not yet set a cluster, use default
+		clusterName = defaultClusterName
+	} else {
+		clusterName, err = clusterSettingStore.GetClusterSetting()
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "Failed to get cluster setting.")
+		}
+	}
 	kurtosisConfig, err := getKurtosisConfig()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting the Kurtosis config")
@@ -104,12 +120,9 @@ func (manager *EngineManager) GetEngineStatus(
 	} else if numRunningEngineContainers == 0 {
 		return EngineStatus_Stopped, nil, "", nil
 	}
-	engineContainer := getFirstEngineFromMap(runningEngineContainers)
 
-	runningEngineIpAndPort := &hostMachineIpAndPort{
-		ipAddr:  engineContainer.GetPublicIPAddress(),
-		portNum: engineContainer.GetPublicGRPCPort().GetNumber(),
-	}
+	// TODO Replace this hacky method of defaulting to localhost:DefaultGrpcPort to get connected to the engine
+	runningEngineIpAndPort := getDefaultKurtosisEngineLocalhostMachineIpAndPort()
 
 	engineClient, clientCloseFunc, err := getEngineClientFromHostMachineIpAndPort(runningEngineIpAndPort)
 	if err != nil {
@@ -221,18 +234,17 @@ func startEngineWithGuarantor(ctx context.Context, currentStatus EngineStatus, e
 	}
 
 	// Final verification to ensure that the engine server is responding
+
 	if _, err := getEngineInfoWithTimeout(ctx, engineClient); err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred connecting to the engine server; this is very strange and likely indicates a bug in the engine itself")
+		return nil, nil, stacktrace.Propagate(err, "An error occurred connecting to the engine server after starting it; "+
+			"if you are running Kurtosis in a Kubernetes cluster, consider running '%v %v' to open a local gateway to the engine running in the cluster", command_str_consts.KurtosisCmdStr, command_str_consts.GatewayCmdStr)
 	}
+
 	return engineClient, clientCloseFunc, nil
 }
 
 func getEngineClientFromHostMachineIpAndPort(hostMachineIpAndPort *hostMachineIpAndPort) (kurtosis_engine_rpc_api_bindings.EngineServiceClient, func() error, error) {
-	url := fmt.Sprintf(
-		"%v:%v",
-		hostMachineIpAndPort.ipAddr.String(),
-		hostMachineIpAndPort.portNum,
-	)
+	url := hostMachineIpAndPort.GetURL()
 	conn, err := grpc.Dial(url, grpc.WithInsecure())
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred dialling Kurtosis engine at URL '%v'", url)